requirements.txt
beautifulsoup4==4.12.2
nltk==3.8.1
pandas==2.1.3
Pillow==10.1.0
rake_nltk==1.0.6
Requests==2.31.0
Wikipedia-API==0.6.0
pyspark==3.5.0